split config tasks in their own files
