+++ smart-contracts/tasks/accounts.js
@@ -1,0 +1,11 @@
+const { task } = require('hardhat/config')
+
+task('accounts', 'Prints the list of accounts', async () => {
+  // eslint-disable-next-line no-undef
+  const accounts = await ethers.getSigners()
+
+  accounts.forEach((account, i) => {
+    // eslint-disable-next-line no-console
+    console.log(`[${i}]: ${account.address}`)
+  })
+})
